Final-project-Nov-24/Dockerized-app/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography==41.0.4

